improvements and testing to HAL, HAL examples

+++ hal/beagleboneblack/include/hal.h
@@ -27,7 +27,7 @@ hal_handle_t BBB_initializeDIOPort(int16_t port, uint8_t dir);
 void BBB_freeDIOPort(hal_handle_t portHandle);
 
 void BBB_setDIO(hal_handle_t portHandle, uint8_t high);
-void BBB_pulseDIO(hal_handle_t portHandle, uint32_t length);
+void BBB_pulseDIO(hal_handle_t portHandle, float length);
 
 uint8_t BBB_getDIO(hal_handle_t portHandle);
